fix(tensilelite): address critical and major review findings in deepseek sgemm
- Signature.py: fix SGPR pad parity by using numStoreSgprNames membership
  check (matching KernelWriter) instead of a partial hand-count; wrong pad
  insertion would bind the wrong 64-bit buffer pointer when UseE, Bias,
  UseScaleCD, or UseScaleAlphaVec were active alongside UseDeepseekScale
- Solution.py: add missing validators in _validateDeepseekScale — reject ISA
  != gfx950, MIArchVgpr=True, GroupedGemm, and StreamK without ForceDPOnly
- ContractionProblem.hpp: change scaleA/B deepseek tensor DataType from Float
  to Int8 to match the E8M0 on-device byte format
- Reference.cpp: add decodeE8M0 helper and use it to read scale bytes
  correctly instead of interpreting them as fp32
- DataInitialization.cpp: initialize deepseek scale buffers with valid E8M0
  byte values rather than fp32 values
- SubtileDeepseekScaleEmit.py: fix nNBlocksStride to use ceil(N/128) instead
  of floor(N/128); use SGPR-space SAddU32+SLShiftRightB32 to avoid illegal
  literal operand in VAddU32; add N=320 to multi-K test shapes to cover
  non-128-aligned stride
- Kernel.py, InstructionEmitter.py: remove dead fp32Scale parameter from
  emitMfmaInstruction and all call sites

+++ projects/hipblaslt/tensilelite/include/Tensile/ContractionProblem.hpp
@@ -354,8 +354,8 @@ namespace TensileLite
             RESIDUAL      = 20, // bf16 input: residual tensor [M_tokens x N_hidden] row-major (optional).
             RSTDBUF       = 21, // f32 input: reciprocal std-dev buffer [M] for RstdScale epilogue.
             QUANTSCALE    = 22, // f32 output: per-tile amax/448 scale [ceil(M/Q0) x ceil(N/Q1)] row-major.
-            SCALEA_DS     = 23, // f32 input: per-row A dequantization scale [M rows].
-            SCALEB_DS     = 24, // f32 input: per-128col-block B dequantization scale [ceil(N/128)].
+            SCALEA_DS     = 23, // E8M0 byte input: per-row A dequantization scale [M rows].
+            SCALEB_DS     = 24, // E8M0 byte input: per-128col-block B dequantization scale [ceil(N/128)].
             TENSOR_COUNT
         };
 
@@ -810,14 +810,14 @@ namespace TensileLite
         {
             if(m_useDeepseekScaleA)
                 m_tensors[TENSOR::SCALEA_DS]
-                    = {"scaleADeepseek", rocisa::DataType::Float, {mRows}, {1}};
+                    = {"scaleADeepseek", rocisa::DataType::Int8, {mRows}, {1}};
         }
 
         void setScaleBDeepseek(size_t nBlocks)
         {
             if(m_useDeepseekScaleB)
                 m_tensors[TENSOR::SCALEB_DS]
-                    = {"scaleBDeepseek", rocisa::DataType::Float, {nBlocks}, {1}};
+                    = {"scaleBDeepseek", rocisa::DataType::Int8, {nBlocks}, {1}};
         }
 
         void setUseBias(int useBias)
